before moving pre execute
